backend/requirements.txt
# Core Flask
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

# Authentication
Flask-JWT-Extended==4.5.3
Flask-Login==0.6.3

# Validation
Flask-Marshmallow==0.15.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# API
Flask-RESTful==0.3.10

# Security
python-dotenv==1.0.0
werkzeug==3.0.0
bcrypt==4.1.1

# Database
SQLAlchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# External APIs
requests==2.31.0
openai==1.3.5

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Production
gunicorn==21.2.0
watchdog==6.0.0